Move some useful things from galsim_test_helpers.py to galsim.test_utils.py
