+++ tests/galsim_test_helpers.py
@@ -24,6 +24,8 @@ import coord
 import copy
 import galsim
 
+from galsim.test_utils import *
+
 
 # This file has some helper functions that are used by tests from multiple files to help
 # avoid code duplication.
@@ -387,246 +389,3 @@ def drawNoise(noise):
     im = galsim.ImageD(10,10)
     im.addNoise(noise)
     return im.array.astype(np.float32).tolist()
-
-
-def do_pickle(obj1, func = lambda x : x, irreprable=False):
-    """Check that the object is picklable.  Also that it has basic == and != functionality.
-    """
-    from numbers import Integral, Real, Complex
-    import pickle
-    import copy
-    # In case the repr uses these:
-    from numpy import array, uint16, uint32, int16, int32, float32, float64, complex64, complex128, ndarray
-    from astropy.units import Unit
-    import astropy.io.fits
-    from distutils.version import LooseVersion
-    from collections.abc import Hashable
-
-    print('Try pickling ',str(obj1))
-
-    #print('pickled obj1 = ',pickle.dumps(obj1))
-    obj2 = pickle.loads(pickle.dumps(obj1))
-    assert obj2 is not obj1
-    #print('obj1 = ',repr(obj1))
-    #print('obj2 = ',repr(obj2))
-    f1 = func(obj1)
-    f2 = func(obj2)
-    #print('func(obj1) = ',repr(f1))
-    #print('func(obj2) = ',repr(f2))
-    assert f1 == f2
-
-    # Check that == works properly if the other thing isn't the same type.
-    assert f1 != object()
-    assert object() != f1
-
-    # Test the hash values are equal for two equivalent objects.
-    if isinstance(obj1, Hashable):
-        #print('hash = ',hash(obj1),hash(obj2))
-        assert hash(obj1) == hash(obj2)
-
-    obj3 = copy.copy(obj1)
-    assert obj3 is not obj1
-    random = hasattr(obj1, 'rng') or isinstance(obj1, galsim.BaseDeviate) or 'rng' in repr(obj1)
-    if not random:  # Things with an rng attribute won't be identical on copy.
-        f3 = func(obj3)
-        assert f3 == f1
-    elif isinstance(obj1, galsim.BaseDeviate):
-        f1 = func(obj1)  # But BaseDeviates will be ok.  Just need to remake f1.
-        f3 = func(obj3)
-        assert f3 == f1
-
-    obj4 = copy.deepcopy(obj1)
-    assert obj4 is not obj1
-    f4 = func(obj4)
-    if random: f1 = func(obj1)
-    #print('func(obj1) = ',repr(f1))
-    #print('func(obj4) = ',repr(f4))
-    assert f4 == f1  # But everything should be identical with deepcopy.
-
-    # Also test that the repr is an accurate representation of the object.
-    # The gold standard is that eval(repr(obj)) == obj.  So check that here as well.
-    # A few objects we don't expect to work this way in GalSim; when testing these, we set the
-    # `irreprable` kwarg to true.  Also, we skip anything with random deviates since these don't
-    # respect the eval/repr roundtrip.
-
-    if not random and not irreprable:
-        # A further complication is that the default numpy print options do not lead to sufficient
-        # precision for the eval string to exactly reproduce the original object, and start
-        # truncating the output for relatively small size arrays.  So we temporarily bump up the
-        # precision and truncation threshold for testing.
-        #print(repr(obj1))
-        with galsim.utilities.printoptions(precision=20, threshold=np.inf):
-            obj5 = eval(repr(obj1))
-        #print('obj1 = ',repr(obj1))
-        #print('obj5 = ',repr(obj5))
-        f5 = func(obj5)
-        #print('f1 = ',f1)
-        #print('f5 = ',f5)
-        assert f5 == f1, "func(obj1) = %r\nfunc(obj5) = %r"%(f1, f5)
-    else:
-        # Even if we're not actually doing the test, still make the repr to check for syntax errors.
-        repr(obj1)
-
-    # Historical note:
-    # We used to have a test here where we perturbed the construction arguments to make sure
-    # that objects that should be different really are different.
-    # However, that used `__getinitargs__`, which we've moved away from using for pickle, so
-    # none of our classes get checked this way anymore.  So we removed this section.
-    # This means that this inequality test has to be done manually via all_obj_diff.
-    # See releases v2.3 or earlier for the old way we did this.
-
-
-def all_obj_diff(objs, check_hash=True):
-    """ Helper function that verifies that each element in `objs` is unique and, if hashable,
-    produces a unique hash."""
-
-    from collections.abc import Hashable
-    from collections import Counter
-    # Check that all objects are unique.
-    # Would like to use `assert len(objs) == len(set(objs))` here, but this requires that the
-    # elements of objs are hashable (and that they have unique hashes!, which is what we're trying
-    # to test!.  So instead, we just loop over all combinations.
-    for i, obji in enumerate(objs):
-        assert obji == obji
-        assert not (obji != obji)
-        # Could probably start the next loop at `i+1`, but we start at 0 for completeness
-        # (and to verify a != b implies b != a)
-        for j, objj in enumerate(objs):
-            if i == j:
-                continue
-            assert obji != objj, ("Found equivalent objects {0} == {1} at indices {2} and {3}"
-                                  .format(obji, objj, i, j))
-
-    if not check_hash:
-        return
-    # Now check that all hashes are unique (if the items are hashable).
-    if not isinstance(objs[0], Hashable):
-        return
-    hashes = [hash(obj) for obj in objs]
-    try:
-        assert len(hashes) == len(set(hashes))
-    except AssertionError as e:
-        for k, v in Counter(hashes).items():
-            if v <= 1:
-                continue
-            print("Found multiple equivalent object hashes:")
-            for i, obj in enumerate(objs):
-                if hash(obj) == k:
-                    print(i, repr(obj))
-        raise e
-
-
-def funcname():
-    import inspect
-    return inspect.stack()[1][3]
-
-
-def timer(f):
-    import functools
-
-    @functools.wraps(f)
-    def f2(*args, **kwargs):
-        import time
-        t0 = time.time()
-        result = f(*args, **kwargs)
-        t1 = time.time()
-        fname = repr(f).split()[1]
-        print('time for %s = %.2f' % (fname, t1-t0))
-        return result
-    return f2
-
-
-class CaptureLog:
-    """A context manager that saves logging output into a string that is accessible for
-    checking in unit tests.
-
-    After exiting the context, the attribute `output` will have the logging output.
-
-    Sample usage:
-
-            >>> with CaptureLog() as cl:
-            ...     cl.logger.info('Do some stuff')
-            >>> assert cl.output == 'Do some stuff'
-
-    """
-    def __init__(self, level=3):
-        from io import StringIO
-
-        logging_levels = { 0: logging.CRITICAL,
-                           1: logging.WARNING,
-                           2: logging.INFO,
-                           3: logging.DEBUG }
-        self.logger = logging.getLogger('CaptureLog')
-        self.logger.setLevel(logging_levels[level])
-        self.stream = StringIO()
-        self.handler = logging.StreamHandler(self.stream)
-        self.logger.addHandler(self.handler)
-
-    def __enter__(self):
-        return self
-
-    def __exit__(self, type, value, traceback):
-        self.handler.flush()
-        self.output = self.stream.getvalue().strip()
-        self.handler.close()
-
-
-# Replicate a small part of the nose package to get the `assert_raises` function/context-manager
-# without relying on nose as a dependency.
-import unittest
-class Dummy(unittest.TestCase):
-    def nop():
-        pass
-_t = Dummy('nop')
-assert_raises = getattr(_t, 'assertRaises')
-if False:
-    # Note: this should work, but at least sometimes it fails with:
-    #    RuntimeError: dictionary changed size during iteration
-    # cf. https://bugs.python.org/issue29620
-    # So just use our own (working) implementation for all Python versions.
-    assert_warns = getattr(_t, 'assertWarns')
-else:
-    from contextlib import contextmanager
-    import warnings
-    @contextmanager
-    def assert_warns_context(wtype):
-        # When used as a context manager
-        with warnings.catch_warnings(record=True) as w:
-            warnings.simplefilter("always")
-            yield w
-        assert len(w) >= 1, "Expected warning %s was not raised."%(wtype)
-        assert issubclass(w[0].category, wtype), \
-                "Warning raised was the wrong type (got %s, expected %s)"%(
-                w[0].category, wtype)
-
-    def assert_warns(wtype, *args, **kwargs):
-        if len(args) == 0:
-            return assert_warns_context(wtype)
-        else:
-            # When used as a regular function
-            func = args[0]
-            args = args[1:]
-            with assert_warns(wtype):
-                res = func(*args, **kwargs)
-            return res
-
-del Dummy
-del _t
-
-# Context to make it easier to profile bits of the code
-class profile:
-    def __init__(self, sortby='tottime', nlines=30):
-        self.sortby = sortby
-        self.nlines = nlines
-
-    def __enter__(self):
-        import cProfile, pstats
-        self.pr = cProfile.Profile()
-        self.pr.enable()
-        return self
-
-    def __exit__(self, type, value, traceback):
-        import pstats
-        self.pr.disable()
-        ps = pstats.Stats(self.pr).sort_stats(self.sortby)
-        ps.print_stats(self.nlines)
